split type defs
